move validators into utils
